requirements.txt
openpyxl
requests
Pillow
streamlit